test: tests combined
